move checker to a separate function
